Move converter to separate class, mark as internal sealed
